+++ Editor/Utility/AssetStoreSamplesInstaller.cs.meta
@@ -1,2 +1,11 @@
 fileFormatVersion: 2
 guid: cd9cd1021d3794a8bb88cc658221e712
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
