+++ Editor/Utility/DependencyInstallerUtilities.cs.meta
@@ -1,2 +1,11 @@
 fileFormatVersion: 2
 guid: 91c165b2af7cd4bdcbf269890f2a49c0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
