+++ Editor/Utility/ItchSamplesInstaller.cs.meta
@@ -1,2 +1,11 @@
 fileFormatVersion: 2
 guid: 0c21c80267e2f4204b7d5e261e616faa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
