+++ Editor/Utility/ManualSamplesInstaller.cs.meta
@@ -1,2 +1,11 @@
 fileFormatVersion: 2
 guid: 23020a8d4675148c9ad05af13ab1202b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
