+++ Editor/Utility/UPMSamplesInstaller.cs.meta
@@ -1,2 +1,11 @@
 fileFormatVersion: 2
 guid: fc1506e515b6046b1939407847cb96e5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
